[#77]/feat deleteDatasInFolders
